rename otl_server => opentelemetry
