--- engine/modules/otl_server/inc/com/centreon/engine/modules/otl_server/otl_data_point.hh
@@ -1,203 +1,0 @@
-/*
-** Copyright 2024 Centreon
-**
-** Licensed under the Apache License, Version 2.0 (the "License");
-** you may not use this file except in compliance with the License.
-** You may obtain a copy of the License at
-**
-**     http://www.apache.org/licenses/LICENSE-2.0
-**
-** Unless required by applicable law or agreed to in writing, software
-** distributed under the License is distributed on an "AS IS" BASIS,
-** WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-** See the License for the specific language governing permissions and
-** limitations under the License.
-**
-** For more information : contact@centreon.com
-*/
-
-#ifndef CCE_MOD_OTL_SERVER_METRIC_HH
-#define CCE_MOD_OTL_SERVER_METRIC_HH
-
-
-namespace com::centreon::engine::modules::otl_server {
-
-/**
- * @brief pair with host_name in first and serv in second
- *
- */
-using host_serv = std::pair<std::string /*host*/, std::string /*serv*/>;
-
-/**
- * @brief This struct is used to lookup in a host_serv indexed container with
- * a std::pair<std::string_view, std::string_view>
- *
- */
-struct host_serv_hash_eq {
-  using is_transparent = void;
-  using host_serv_string_view = std::pair<std::string_view, std::string_view>;
-
-  size_t operator()(const host_serv& to_hash) const {
-    return absl::Hash<host_serv>()(to_hash);
-  }
-  size_t operator()(const host_serv_string_view& to_hash) const {
-    return absl::Hash<host_serv_string_view>()(to_hash);
-  }
-
-  bool operator()(const host_serv& left, const host_serv& right) const {
-    return left == right;
-  }
-  bool operator()(const host_serv& left,
-                  const host_serv_string_view& right) const {
-    return left.first == right.first && left.second == right.second;
-  }
-  bool operator()(const host_serv_string_view& left,
-                  const host_serv& right) const {
-    return left.first == right.first && left.second == right.second;
-  }
-  bool operator()(const host_serv_string_view& left,
-                  const host_serv_string_view& right) const {
-    return left == right;
-  }
-};
-
-using metric_request_ptr =
-    std::shared_ptr<::opentelemetry::proto::collector::metrics::v1::
-                        ExportMetricsServiceRequest>;
-
-/**
- * @brief some metrics will be computed and other not
- * This bean represents a DataPoint, it embeds all ExportMetricsServiceRequest
- * in order to avoid useless copies
- *
- */
-class data_point {
- public:
-  enum class data_point_type {
-    number,
-    histogram,
-    exponential_histogram,
-    summary
-  };
-
- private:
-  const metric_request_ptr _parent;
-  const ::opentelemetry::proto::resource::v1::Resource& _resource;
-  const ::opentelemetry::proto::common::v1::InstrumentationScope& _scope;
-  const ::opentelemetry::proto::metrics::v1::Metric& _metric;
-  const google::protobuf::Message& _data_point;
-  const ::google::protobuf::RepeatedPtrField<
-      ::opentelemetry::proto::common::v1::KeyValue>& _data_point_attributes;
-  uint64_t _nano_timestamp;
-  data_point_type _type;
-  double _value;
-
-  data_point(
-      const metric_request_ptr& parent,
-      const ::opentelemetry::proto::resource::v1::Resource& resource,
-      const ::opentelemetry::proto::common::v1::InstrumentationScope& scope,
-      const ::opentelemetry::proto::metrics::v1::Metric& metric,
-      const ::opentelemetry::proto::metrics::v1::NumberDataPoint& data_pt);
-
-  data_point(
-      const metric_request_ptr& parent,
-      const ::opentelemetry::proto::resource::v1::Resource& resource,
-      const ::opentelemetry::proto::common::v1::InstrumentationScope& scope,
-      const ::opentelemetry::proto::metrics::v1::Metric& metric,
-      const ::opentelemetry::proto::metrics::v1::HistogramDataPoint& data_pt);
-
-  data_point(
-      const metric_request_ptr& parent,
-      const ::opentelemetry::proto::resource::v1::Resource& resource,
-      const ::opentelemetry::proto::common::v1::InstrumentationScope& scope,
-      const ::opentelemetry::proto::metrics::v1::Metric& metric,
-      const ::opentelemetry::proto::metrics::v1::ExponentialHistogramDataPoint&
-          data_pt);
-
-  data_point(
-      const metric_request_ptr& parent,
-      const ::opentelemetry::proto::resource::v1::Resource& resource,
-      const ::opentelemetry::proto::common::v1::InstrumentationScope& scope,
-      const ::opentelemetry::proto::metrics::v1::Metric& metric,
-      const ::opentelemetry::proto::metrics::v1::SummaryDataPoint& data_pt);
-
- public:
-  const ::opentelemetry::proto::resource::v1::Resource& get_resource() const {
-    return _resource;
-  }
-
-  const ::opentelemetry::proto::common::v1::InstrumentationScope& get_scope()
-      const {
-    return _scope;
-  }
-
-  const ::opentelemetry::proto::metrics::v1::Metric& get_metric() const {
-    return _metric;
-  }
-
-  const google::protobuf::Message& get_data_point() const {
-    return _data_point;
-  }
-
-  uint64_t get_nano_timestamp() const { return _nano_timestamp; }
-
-  data_point_type get_type() { return _type; }
-
-  const ::google::protobuf::RepeatedPtrField<
-      ::opentelemetry::proto::common::v1::KeyValue>&
-  get_data_point_attributes() const {
-    return _data_point_attributes;
-  }
-
-  double get_value() const { return _value; }
-
-  template <typename data_point_handler>
-  static void extract_data_points(const metric_request_ptr& metrics,
-                                  data_point_handler&& handler);
-};
-
-template <typename data_point_handler>
-void data_point::extract_data_points(const metric_request_ptr& metrics,
-                                     data_point_handler&& handler) {
-  using namespace ::opentelemetry::proto::metrics::v1;
-  for (const ResourceMetrics& resource_metric : metrics->resource_metrics()) {
-    const ::opentelemetry::proto::resource::v1::Resource& resource =
-        resource_metric.resource();
-    for (const ScopeMetrics& scope_metrics : resource_metric.scope_metrics()) {
-      const ::opentelemetry::proto::common::v1::InstrumentationScope& scope =
-          scope_metrics.scope();
-
-      for (const Metric& pb_metric : scope_metrics.metrics()) {
-        switch (pb_metric.data_case()) {
-          case Metric::kGauge:
-            for (const NumberDataPoint& iter : pb_metric.gauge().data_points())
-              handler(data_point(metrics, resource, scope, pb_metric, iter));
-            break;
-          case Metric::kSum:
-            for (const NumberDataPoint& iter : pb_metric.sum().data_points())
-              handler(data_point(metrics, resource, scope, pb_metric, iter));
-            break;
-          case Metric::kHistogram:
-            for (const HistogramDataPoint& iter :
-                 pb_metric.histogram().data_points())
-              handler(data_point(metrics, resource, scope, pb_metric, iter));
-            break;
-          case Metric::kExponentialHistogram:
-            for (const ExponentialHistogramDataPoint& iter :
-                 pb_metric.exponential_histogram().data_points())
-              handler(data_point(metrics, resource, scope, pb_metric, iter));
-            break;
-          case Metric::kSummary:
-            for (const SummaryDataPoint& iter :
-                 pb_metric.summary().data_points())
-              handler(data_point(metrics, resource, scope, pb_metric, iter));
-            break;
-        }
-      }
-    }
-  }
-}
-
-};  // namespace com::centreon::engine::modules::otl_server
-
-#endif
