--- packaging/centreon-engine-otel-listener-debuginfo.yaml
@@ -1,42 +1,0 @@
-name: "centreon-engine-otel-listener-debuginfo"
-arch: "${ARCH}"
-platform: "linux"
-version_schema: "none"
-version: "${VERSION}"
-release: "${RELEASE}${DIST}"
-section: "default"
-priority: "optional"
-maintainer: "Centreon <contact@centreon.com>"
-description: |
-  Debuginfo package for centreon-engine.
-  Commit: @COMMIT_HASH@
-vendor: "Centreon"
-homepage: "https://www.centreon.com"
-license: "Apache-2.0"
-
-contents:
-  - src: "../lib/libotl_server.so.debug"
-    dst: "/usr/lib/debug/usr/lib64/centreon-engine/libotl_server.so.debug"
-    file_info:
-      mode: 0644
-
-overrides:
-  rpm:
-    depends:
-      - centreon-engine-otel-listener = ${VERSION}-${RELEASE}${DIST}
-  deb:
-    depends:
-      - centreon-engine-otel-listener (= ${VERSION}-${RELEASE}${DIST})
-    conflicts:
-      - centreon-engine-otel-listener-dbgsym
-    replaces:
-      - centreon-engine-otel-listener-dbgsym
-    provides:
-      - centreon-engine-otel-listener-dbgsym
-
-rpm:
-  summary: Debuginfo package for centreon-engine-otel-listener.
-  compression: zstd
-  signature:
-    key_file: ${RPM_SIGNING_KEY_FILE}
-    key_id: ${RPM_SIGNING_KEY_ID}
