--- packaging/centreon-engine-otel-listener.yaml
@@ -1,34 +1,0 @@
-name: "centreon-engine-otel-listener"
-arch: "${ARCH}"
-platform: "linux"
-version_schema: "none"
-version: "${VERSION}"
-release: "${RELEASE}${DIST}"
-section: "default"
-priority: "optional"
-maintainer: "Centreon <contact@centreon.com>"
-description: |
-  Centreon Engine OpenTelemetry listener.
-  Commit: @COMMIT_HASH@
-vendor: "Centreon"
-homepage: "https://www.centreon.com"
-license: "Apache-2.0"
-
-contents:
-  - src: "../lib/libotl_server.so"
-    dst: "/usr/lib64/centreon-engine/libotl_server.so"
-
-overrides:
-  rpm:
-    depends:
-      - centreon-engine-daemon = ${VERSION}-${RELEASE}${DIST}
-  deb:
-    depends:
-      - centreon-engine-daemon (= ${VERSION}-${RELEASE}${DIST})
-
-rpm:
-  summary: Centreon Engine OpenTelemetry listener.
-  compression: zstd
-  signature:
-    key_file: ${RPM_SIGNING_KEY_FILE}
-    key_id: ${RPM_SIGNING_KEY_ID}
